Gameplay Effect Delegates
Implemented a delegate to notify when a gameplay effect is applied to an ability system component
Created EffectApplied callback in ASC. Made the AuraCharacter::InitAbilityActorInfo into a virtual in the CharacterBase and created overrides in character and enemy. The Init override functions will now cast to the ASC and call the callback and bind the delegate, broadcasting pointer for the the ASC, const ref for the effectspec and the effect handled.
For testing added an onscreen message and a breakpoint to the callback